Fixed BgkCollisions to be MPI-SHM-ed. Small cleanup of FiveMomentSrc
Added regression test for Eigen based semi-implicit solver. Usually we
will use the exact solver (as it is 10x faster), but as we have the
Eigen solver it is best to test it. (Eigen based solver is easier to
mod with new source terms as it creates a matrix to invert, while
exact source solver with new terms may not be possible to construct).