A species object does not have initField() method! Instead, one needs initDist() method. Fixed this in FuncVlasovSpecies
Fixed a bug in KineticSpecies that was only triggered if a species was
not evolved. This crept in when the diagnosticMomentFields was indexed
with moment name rather than index.

Added a regression test for FuncVlasocSpecies based on Buneman
instability. Results look almost exactly the same as "ghost species"
test in my SimJ. Small differences to exist, but these are as moments
are not exact. I think! However, growth rates look exactly the same.