Support for spatially varying, cell-wise constant collisionality in VmLBO. Had to revert back to LBO that takes u and vtSq instead of nuU and nuVtSq. Created an updater, SpitzerCollisionality, to rescale the collisionality at every time step. Also tried rescaling without an updater in the equation object but decided updater rocks more, especially in the future.
Regression tests are likely to fail after this. But I visually inspected
most of the lboRelax tests and they seem fine. The conservation tests
fail but due to machine precision differences (order of operations has
changed).

In my mac I'm having a problem where I can't runregression.lua run check
-r vm-lbo because it hangs at vm-lbo/rt-lboRelax-1x1v-p2.lua. But if you
run check just this test it passes.

Also, there isn't yet a goot regression test for varNu. There is
Regression/vm-lbo/rt-lboRelaxVarNu-1x1v-p1.lua, but that is just to test
infrastructure (density and temperature are constant therein).

I'm also commiting maxima scripts I started for the case of
collisionality varying within a cell, but that challenge is not being
pursued atm. Nu would have to be included in SelfPrimMom, but more
puzzling is how the computation of 1/T^(3/2) ought to be done. Fun.