Add support for spatially varying, cell-wise constant collisionality in GkLBO. Changed GkLBO kernels to take uPar and vtSq instead of nuU and nuVtSq. GkLBO app uses SpitzerCollisionality to compute spatially varying nu when desired.
Adding an input file to test the infrastructure of spatially varying nu
in GkLBO to the Regression/gk-lbo folder, but it is not a good test to
check if the implementation is physically correct. Still need to come up
with one.

Made a minor change to Eq/VmLBO.lua to avoid errors in the occasion that
the SelfPrimMom limit crossing is found in the volume update method
(slightly redundant because selfPrimMom is velocity-space independent so
it should be found in volume, surface and boundarySurf updates).