FluidSpecies: No default name. Otherwise the one given in App is overwritten.
Is it intended for a fluid species to have a default name "name"? It would be
good to do that, but then the KineticSpecies should have a similar behavor,
too, and setName should be called after fullInit (but KineticSpecies seems to
need a name in its fullInit).

Overall, it might be nice to clean up and unify the interfaces of different
species.