Add (optional) diffusion term in the Incompressible Euler equation.
For now, it work like collisions in the kinetic solvers. One has to
specify a diffusion table within each species, and a list of diffusion
coefficients (in each direction).

See Regression/positivity/rt-pos-diffAdv.lua

In the future we could consider generalizing this to an App that adds
configuration and/or velocity space diffusion in kinetic equations as
well.