In the binOp reorganization, polyOrder 3 was deleted. Since I am using polyOrder 3 for a number of simulations I have regenerated the Serendipity basis, polyOrder 3 data for 1X1V through 2X2V. There also appeared to be a bug in the maxima script for the weak divide.
When generating the weak divide for phase space data, the command to invert the matrix was: data->u_D = data->AEM_D.colPivHouseholderQr().solve(data->BEV_D);

However, when generating the weak divide for configuration space data, the command in the maxima script was: data->u_S = data->AEM_S->colPivHouseholderQr().solve(data->BEV_S);

The latter command appears to be incorrect because AEM_S is not a pointer. I have fixed this in the maxima script. To be honest, I am not sure how this error propagated into my pull of the code when the polyOrder 1 and 2 generated code was correct.