A bug creeped into the shared memory functionality on 11/21/18, changeset 3cef75a, when the method of computing the CFL was improved to avoid failed time-steps. To make this work, applyBc was moved into the combineRK method, but the accumulate method and applyBc have different loop structure so boundary conditions were being applied to some regions of the domain before the Runge-Kutta step had finished being combined (some processes were outrunning others on the accumulate part but then treating different parts of the domain when applying boundary conditions). Two-stream instability now gives the same results on 4 cores shared vs. non-shared memory. More testing is needed but this seems to have been the bug.
I am also anticipating a bug when the GK algorithm becomes shared memory compatible and I have added two barriers over the shared communicator when only applying the boundary conditions to the fluctuations. This may not be needed but I did not want the GK team to have to track this down as we extend shared memory compatibility.

I have also changed a few iterators to row-major to be in line with the recent reorganization. I don't think this does anything because how we loop through the domain shouldn't really matter, but just to be consistent.