Added an "examples" tool to the system. Now one can get in-build examples from installed Gkeyll package, including description of the example. One hopes there will be many good examples added to the system soon. A good example should be easy to run (though unlike a regression test it could take longer or need parallel resources to run) and should come with a brief description.
The description should tell the user what the example does, how to
make some typical plots, what resources are needed to run (specially
if this is a parallel simulation) and other things. The description
should be in markdown format. Equations can be included. Some users
may want to pipe the description to pandoc and save the output as PDF
or Word document and would hence would get nicely rendered equations.