Making the BGK operator shared memory safe. This involved adding the appropriate barriers to the BGK App and making the BGK updater and MaxwellianOnBasis updaters shared memory safe. Some further refactoring may be necessary after the improvements to cross-collisions for the LBO are imported to the BGK App, but this works for now.
I have further done a not insignificant amount of code clean up on a number of updaters. A few updaters had never imported functionality we added to the code, such as the grid:getDx() method that's marginally faster. Further, many updaters were initializing an over-abundance of indexers (mostly my fault) when in fact in the process of making the MaxwellianOnBasis and LagrangeFix updaters shared memory safe I realized you only need two indexers: a configuration space indexer and a phase space indexer. I have noticed small performance boosts on my laptop but these could be within the noise of the machine. Regardless it's almost certainly better to not be initializing so many indexers, especially with how LuaJIT's garbage collector behaves when it's managing a lot of small chunks of memory.

I also went through PlasmaOnCartGrid and KineticSpecies and cleaned up some of the MPI barriers since many of them are not necessary with the movement of the barrier to inside the sync() method.

I have also tried to standardize a few names for things across updaters, such as the name of indexers. This is aesthetic, and I'm agnostic on what the best names are, I just thought while I was cleaning up some of the code I would try to standardize this. If you hate the names you can change them all, not a big deal to me. But I think some of the updaters are more readable now.