For now settle for the following Dirichlet and Neumann BCs for the diffusion term (p=1).
Dirichlet:
Construct a recovery and require that it matches the Dirichlet value at
the boundary, and require the ghost cell higher moments to be the same
as the skin cell moments.

Neumann:
Construct a recovery and require that it matches the Neumann value at
the boundary, and that it matches the value of the skin cell solution at
the boundary.