Implement Dirichlet and Neumann BCs for constant diffusion terms using .cpp kernels. Also, the user is expected to indicate the desired Dirichlet/Neumann value in the input file, for example:
bcx = {{Plasma.IncompEulerSpecies.bcDirichlet, 1.0}, {Plasma.IncompEulerSpecies.bcNeumann,2.0*math.pi-1.0}},

The Dirichlet condition is implemented by constructing a maximal-order
recovery polynomial across the skin-ghost cells and requiring that it
matches the Dirichlet value at the boundary, and that the higher
even/odd moments have odd/even symmetry.

The Neumann condition is implemented by constructing a maximal-order
recovery polynomial across the skin-ghost cells, and requiring that it
has all the same derivatives (including the zeroth derivative) at the
boundary as a polynomial constructed in the skin cell with one more
degree of freedom (used to match the Neumann value there).

This Neumann option is slightly inferior in the 1x, p=1 vanLeer 2005
case compared to building a maximal-order polynomial across skin-ghost
cells and requiring that it matches the Neumann derivative value at the
boundary, and that it matches the value of the function in the ghost
cell at the boundary. But it is not clear how to generalize this to
higher polyOrder or dimensionality.