Removed the runregression.lua file and now installed it as part of the Gkeyll Tool system. To run regression tests you need to be in the gkyl/Regression directory and run it like:
gkyl runregression .....

Note that the .lua extension has vanished. This may appear like a cosmetic change but it allows the regression testing to be treated like any other Tool, which is shipped as part of the Gkeyll install. At present, this Tool can't be run without the source checkout, but I will fix this so at least some tests can be run by users who did not build the code from source and only have access to an installed binary distribution.