DG interpolation updater
Add a new updater that transfers a cartesian grid from one grid to another. So far it has been tested in 1D and 2D and appears to work (more diagnostics needed), but the code is such that it hopefully works for higher dimensions as well. Before going to higher dimensions it would be good to:
1) Diagnose/test 1D and 2D more.
2) Find ways to simplify/optimize the kernels. They are going to be huge and ugly at higher D otherwise.

Current limitations:
- Transfer between two (whole) grids for the same domain. Not yet capable of
  doing grid subsets.
- It may only work for (2^a)*(3^b) grids, i.e. number of cells factorizes into
  powers of 2 and 3. It is hard to come up with simple rules for indices and stencil sizes for arbitrary grids.