Adios group name change for CartField IO
Recent changes make it so that when species had the same number of cells, the same attribute values (e.g. for lower/upper bounds) were written in files of different species, even if these had different values for the species.

so: if you run a simulation were electrons and ions have the same number of cells, but different velocity limits, the attributes lowerBounds/upperBounds in the electron and ion files had the same values.

IMPORTANT: it seems that unit/regression tests did not catch this? perhaps we should check whether unit/regression tests are checking attribute values.