Serious warning: Often we do things like:
e : doExpand(r, basis)$

Now, this is okay is r is an atom or a list. HOWEVER, if r is an expression then the doExpand method will not work as expected and trouble will ensue. If you use doExpand() inside your own function please declare 'r' as a local variable:

foo() := block([r],
  ....
  e : doExpand(r,basis)
  ...
  )$

Now, even if we have something like:

r : x^2$
foo()

inside foo() the "r" that appears in doExpand() will not be x^2 but will be the symbol 'r as expected. I will add an error message to doExpand to ensure this never happens.