HyperDisCont cuda kernel is now working and producing correct results for the Unit/test_HyperDisContCuda.lua test, which directly compares the results of a host and device call of HyperDisCont. caveats are that the HyperDisCont kernel is hard-coded to assume the equation object is from Vlasov class, and the Vlasov class is hard-coded to assume a particular dimensionality/polyOrder/basisType. nonetheless, this is a significant step, and abstractions to make it more flexible will hopefully be coming soon. there are also a few important aspects of HyperDisCont that have not yet been implemented in the kernel (e.g. cfl calculation).
needed to get rid of some of the cuda shared library linking and fall back to linking cuda object files from Eq and Updater into a single culink.o. this is because nvcc cannot do device-code linking across a shared library boudnary. perhaps there is a better solution, but for now this works (I think... there were some issues with the kerneldriver linking but hopefully it is ok now.)