Working in testing ideas for abstracting out the equation object. Unfortunately, it is not yet working, though seems the idea is sound. For some reason the cudaMemcpyFromSymbol is failing (returning error code 13 which means "ErrorInvalidSymbol"). Not sure why this should be the case as the function I am trying to copy from device -> host is a device function. Cuda is really frustrating.
Added a bunch of new error codes and their names to the cuda Runtime. Added string representations of errors.

Updates to cuda test cases.

Working with C structs is rather tedious at present. I will add an abstraction to allow taking a C-struct and just copying it to device without the tedium of all the various cuda calls (and also hence prevent memory leaks by hooking in the free calls to the __gc function). This will make working with C-structs (alloc-ing them, setting them and copying them to device) much more ppleasant. Of course, CUDA is highly unpleasant to start with. A royal PITA.

Need to add some C++ stuff to work with cuda types also. Anyway, all fun stuff ... getting the hang of it.