Added proper min/max types for reduce function. Now reduce should work with int, float, double, long int.
Found another place where a meta-object (this time an allocator) was being created everytime a CartField is being created. However, this should only happen once for each *type* stored in CartField, i.e. for all CartFields that store double only *one* allocator is needed etc.

There is still a serious issue here with MPI when the CartField stores a non-numeric type (like a struct). Currently I think only part of the object will be sync()-ed in this. Will add an issue on this.

Further, the present implementation makes another error in terms of assuming all numeric types are doubles and CUDA fields are also doubles. Not sure how to generalize.

Overall, the CartField is seriously complicated (though only has < 1000 lines of code). Need to refactor by splitting out the sync() and several other methods