Please read:
This is a breaking change for ALL input file. Basically, every input file needs to be modified slightly (or more if you used old-style of require "App.PlasmaOnCartGrid"). So if you had:

local Plasma = require ("App.PlasmaOnCartGrid").VlasovMaxwell

replace it by:

local Plasma = require ("App.PlasmaOnCartGrid").VlasovMaxwell()

Similarly for Gyrokinetic, Moment etc App sub-systems.

Why was this change needed?

In the past we were loading the whole of Gkeyll's various solvers and kernels into the App system. This is a bad idea: say one is doing a GK sim. Why should the VM, Moments, IncompEuler, etc etc etc stuff be loaded? The serious consequence of this was that as have 1000s of kernels at this point (> 900 kernels cpp files, each with many kernels) the ffi.cdef system's internal table to hold the external C functions was getting flooded and was leading to a table overflow error.

Going forward, as part of the refactor I envision over the next few months I want to get to a point in which we ONLY ffi.cdef those kernels call we needed. We can do this but it needs work. Before that we need to split the kernels out into their own repo, clean up the mess accumulated in various infrastructure things (JJ and I have discivered some peculiar behaviour of the CartField, for example).