First attempt at a function to apply periodic boundary conditions with a simple copy with Mpi.comm_size(nodeComm) = 1 (in other words we are performing a simulation on a single compute node with either MPI SHM, or a single GPU).
I do not know yet whether the GPU version works (committing to pull on Adroit and test). Issues I forsee: I am fetching the appropriate ranges (skin cells and ghost cells) in Lua and then passing them to the device function. Is this legal? If it is not, what is the best way to do the function call decomposedRange:subDomain(1):lowerSkin(dir, self._upperGhost) (and similar calls) to get the appropriate sub ranges to be looped over for the copy?

Other annoying thing to fix: to make this copy shared memory safe with MPI SHM, I have a temporary buffer so that I can loop over the two ranges separately (loop over skin cells -> copy -> loop over ghost cells -> put copy in appropriate place). Although the GPU version still needs to be debugged, you can see that it is a much cleaner implementation as I can simply get the appropriate index from the inverse indexer on the two different ranges. While I have examples from HyperDisCont for how to do this with the GPU implementation, I am not sure how to do such a thing when MPI SHM is turned on. We create the decomposition to loop over ourselves when MPI SHM is turned on, so I am not sure how easy/hard it is to break up these sub ranges inside a C function, call the inverse indexers, and make the appropriate copies.

This is not likely to be a significant performance bottleneck either way, but it would clean up CartField if I did not need different functionality depending on whether I am running on the CPU/GPU.