Fixing a bunch of errors in CartField from the implementation of the devicePeriodicCopy method.
I need to define GkylCartField_t before the signature for devicePeriodicCopy, so the initial C interface has been rearranged slightly.

Also fixes to the signature so that a CartField is actually passed to the method.

There was an error in the initialization that I was checking the size of nodeComm, but this is not defined when running with MPI SHM. This has been fixed.

NOTE: This method still does not work because I cannot seem to pass the ranges to the device method the way I am currently doing it, but the code at least runs and tells me the unit test failed (as opposed to not even running due to a runtime error).