Geo file w/ multiple fields + read it if given
a) Now the allGeo file has a separate dataset for each geometry-related
  quantity. This makes the file (more) self-describing.
b) Because of (a) maybe there is no point to writing separate files for the metric quantities in the write: method anymore.
c) Now the input file can specify the allGeo file through allGeoFile in GkGeometry. This way it just reads the geo quantities instead of computing them again.

Option (c) saves time, which helps when initialization takes hours. However most of the time is spent in computing ICs, sources and other projected quantities. It appears to me that the only way to truly save initialization time is to allow each of these operations (e.g. source projection) to read in a file so we don't have to compute them again if they are to be the same as a previous simulation.