Option to read in projection objects
Now we can use keyword fromFile to read projection objects (e.g. background, sources) from a file. This speeds up initialization if such quantities are the same as from a previous simulation and remained the same (some scaling with power and M012 may still occur).

For a simulation input file "inpuFile.lua" which wrote a source for species "elc" in a file named "inputFile_elc_fSource_0.bp", we can read in the source by placing

fromFile="elc_fSource_0.bp"

in the source table. This makes me realize that we should allow restarts to output a new source file with the latest frame name (i.e. not "_0.bp") for the cases that we restart an earlier simulation but with different source/power (e.g. LH transition?).

Regression tests: VM tests passed. gk-kbm fails only because the background and sources are now written with ghost cells. Could not run other gk tests because there are some kernels missing in this branch. This makes me think that we should merge master in here soon.