world option for MappedGrid
One can now pass

world = {dim = wD, evaluateAt = evalTbl}

in the plasma table to specify the higher dimensional world that
needs to be considered when computing the metric. Here `wD` is the
dimensionality of that world, and evalTbl is a key-value pairs table
containing the coordinates of the non-simulated dimensions. So for example
for a 1x2v simulation that requires 3D space to compute the metric we may pass

world = {dim=3, evaluateAt={x=1.1, y=0.0}}

GK reg tests with geo still pass. Jacobian and magnetic field in a 1x2v
sim with geometry appear to be correct. However this 1x2v simulation when
allowed to run eventually encounters a seg fault.

Also, some 1x2v simulations that run in master do not run here. Perhaps it's
time to consolidate the two branches.