Correct FEM projection kernels
- For some reason we were double counting in 2D and triple counting in 3D when projecting onto the FEM basis.
- Subtract the volume average from the source before projection, although it turns out this doesn't seem to matter because subtraction of a constant and projection commute.

This commit corrects an error in which 2D periodic solves exhibitted the right convergence but arrived at the wrong solution. We hadn't catched this error because I didn't check the accuracy of the 2D solutions, only the convergence. This problem also went unnoticed because FEM projection doesn't yet have a regression test.