A commit to mark GK_gen_geo -> master merge
This commit just deletes a commented line of code, but it's really to explicityly say in this comment that the previous commit brought in GK_gen_geo via renameGeoKernels. In order to merge GK_gen_geo the work flow was this one:

1. In branch mergeMaterAndGKGeo we renamed the GK kernels in master (and adapted the infrastructure).
2. In branch renameGeoKernels we renamed the GK kernels in GK_gen_geo (and adapted the infrastructure).
3. Merged mergeMaterAndGKGeo -> renameGeoKernels.
4. Merged master -> mergeMaterAndGKGeo.
5. Again merged mergeMaterAndGKGeo -> renameGeoKernels.
6. Finally merged renameGeoKernels -> master.

We intend to delete mergeMaterAndGKGeo and renameGeoKernels, but will keep GK_gen_geo to preserve the history of gen_geo development.

Thank god for regression and unit tests.