Modification of the zero-flux boundary condition when running with hyperdiffusion. Following the logic in the LBO, where our zero-flux boundary condition sets df/dv = 0 at the wall, we set all odd derivatives at the wall to be zero like we are applying the diffusion operator multiple times (so df/dv and d^3f/dv^3 = 0 for grad^4, and additionally d^5 f/dv^5 for grad^6). This new boundary condition further assists in stabilizing Vlasov-Maxwell simulations that are unstable when run collisionlessly.
This commit also adds the ability to run with grad^4 with p=1, since the recovery for p=1 is enough to generate a cubic and allow for the solution of grad^4. In general, I am finding that grad^6 with p=2 is better than grad^4 with either p=1 or p=2, and that tensor bases provide further robustness improvements. More testing is needed, but for now we are moving in the right direction.

Will begin work on checking grad^4 and grad^6 in 2V and 3V simulations to see if they stabilize reconnection and collisionless shock simulations which have been crashing for Jason and I.