Adding 3D and 4D diffusion kernels now that we have converged to a sensible boundary condition. Preliminary tests suggest 2x2v simulations with velocity space hyperdiffusion are more stable and some convergence metrics look just as good (field energy still collapsing as it should in Skoutnev et al. simulation setup).
Also adding the higher dimensional kernels for running with grad^4 and p=1. Need to extend diffusion kernels to at least 5D so we can use them with the reconnection simulations we have been attempting.

Perhaps need to think of even more tests to run and benchmark these kernels. Right now things look good but the higher dimensional tests are expensive to benchmark with..