Upgrading the default version of openmpi
Unless another MPI library was specified (as commonly done in clusters), we had been building gkyl with openmpi-3.1.2. We found a bug in openmpi-3.1.2 which caused a segmentation fault when MPI_Comm_create and MPI_Cart_sub were used at the same time. This and potentially other communicator creation-related bugs are not due to the way we wrap MPI in gkyl; one can write a C MWE to show this.

This bug appears to be fixed in the latest stable version openmpi-4.0.5, so we are upgrading the gkyl build to default to this version. We have tested gkyl builds with openmpi-4.0.5 on MacOS Mojave and Catalina, and on the MIT Engaging cluster. Such tests included:
Unit/test_Mpi.lua
Unit/test_CartGridPar.lua
A 3x2v/3x3v gyrokinetic/vlasov simulation with neutrals.

Note that the above bug does not exist in all MPI distributions; Intel MPI 2019, for example, does not appear to have such bug. We are also not presently using MPI_Cart_sub, but may begin using it soon. Therefore there's no immediate risk of continuing to use the MPI you already have. But we advise that you upgrade as soon as it is convenient for you.

I will send out an email about this in the coming days, and perhaps again when I begin using MPI_Cart in a way that may affect others.