Twist-shift improvement & prep for 3x2v
Using a step function as the donor field and a const yShift I realized that the TwistShift code is not doing the projections (integrals) in the shifted region but rather in the donor region that overlaps with the shifted region (there's a subtle difference). This is easy to fix for scenario sxv-sxvi integrals, but no so easy for the others. I think this commit fixes it for sxv-sxvi and gives better results. Scenarios i-iv I think can be fixed the same way if we re-write them as integrals with variable y-limits instead of variable x-limits. The other scenarios however I don't yet know how to improve.

I also began prepping for applying the BC to the z-ends of a 3D field (or 3x2v field). Not yet ready.