New way of looping over species table
Previously we had been looping over the species tables with loops like

for nm, s in pairs(species) do
   ...
end

`pairs` is the correct iterator for a table of name-value pairs. However, it is not deterministic: there is no guarantee that consecutive runs will loop in the same order. More importantly, there's no guarantee that different MPI ranks will loop in the same order. This is a problem for I/O (and possibly other operations), as one rank may be trying to output the ``ion`` distribution function at the same time as another rank may be outputting the ``electron`` distribution function.

To remedy this I've replaced all the species loops I could find with loops like

for i = 1, #species["keys"] do
   local nm, s = species["keys"][i], species[species["keys"][i]]

   ...
end

`keys` is a table of sorted key names which will be identical in every MPI rank and resides inside of species. This is the way we should loop over species from now on.

It seems a little cumbersome compared to the elegant `pairs(table)` approach. So two other options we could consider are:
- storing `keys` as the metatable of species.
- creating an iterator in a central location that works like `pairs` but loops
  in the order given by `keys`.