A few bug fixes on the gradient-based closure. The temporary array used to store the symmetrized heat flux had the wrong number of components. This is fixed now. In addition, there should have been a factor of 1/3 in the computation of the symmetrized heat flux. With these changes, the regression test which uses the gradient based heat flux is able to stably run (even without super-time-stepping) at values of the "conductivity" that are reported in Jonathan Ng's 2020 PoP paper (the length scale the heat flux should most effectively conduct is de for electrons and di for ions is the argument for what the parameter alpha should be within the closure).
Whether the implementation is 100 percent correct remains to be determined. The results look a bit strange at coarse resolutions, but at higher resolutions run on my desktop, the reconnection layer starts to look more reasonable. It is possible the integration in time is still not as accurate as it could be and the super-time-stepping method should be ported over into the update method for fluids such that the integration in time goes:
sources(dt/2) hyperbolic(dt) diffusion(dt) sources(dt/2)

with the diffusion step of size dt handled with super-time-stepping. I would feel a lot better about the comparison if we revamped this part of the fluid App to allow this (and this could interface nicely with Liang's additions which allow for the addition of viscosity).