Getting closer. I realized there were two mistakes in the indexing within the updater: the first was that idxl and idxr needed to be copied within the loop over directions, or else the stencil gradually moved diagnoally (not ideal). The second was that the first loop over the grid to compute grad(T) and q needs to be extended by one cell (into the first ghost cell layer) so that div(Q) has all the information it needs (essentially we still need a five-point stencil). Fortunately the limiters already require that ghost = {2,2} so all the information is in place.
Now if I do a decomposition in only the periodic direction, from one core to eight cores, the answer remains the same. However, the answer is still changing when I do a decomposition in the non-periodic direction, suggesting that the boundary conditions in non-periodic directions are not being applied correctly.