AnisotropicDiffusion: calcQ and calcDivQ switches.
They can be changed. Combined with the updateTemperature option, a user
can perform calcQ, calcDivQ, and updateTemperature steps separately,
thus be able to further process Q or divQ before the next step. For
example, we may add additional terms to Q, like in the Braginskii heat
conduction, the additional parallel heat to the electrons.