Reorganize DistFuncMomentCalc + change its initialization
The :advance method of DistFuncMomentCalc was horrible and hard to follow. Too many branching statements had to be evaluated within phase-space and time loops.

Split the advance method into various kinds for each of the specific scenarios needed: single moment, five/three moment and LBO five/three moment calculations. Then during initialization we pick the right advance method that will be used in the time loop. This does away with nearly every branching statement that was there before.

Also change the way to initialize DistFuncMomentCalc so that now we have to pass sample fields that we would pass to the advance method (e.g. distf and numDensity). This way we can pre-define ranges and other objects that are used in the advance method.