Now that the GK+neutrals simulations are testing p=2 and there is thus a need for certain p=2 kernels, the cas-scripts are adjusted so that these can be generated. Small change mostly (not setting maxPolyOrder to 1 if generating the 6D kernels) and this change is limited to the CAS scripts that are required *only* for neutrals (streaming, moment calculation, and primitive moments).
Also the streaming surface kernels are now structured similarly to kernels such as LorentzFuncs (for the Vlasov-Maxwell acceleration) so that Ghat is computed on the surface basis, and then the corresponding update is generated. This makes virtually no difference in cost/size of kernel in low dimensionality, but the higher dimensional kernels (especially p = 2) are now 20% smaller suggesting an improvement. If nothing else, so long as the cost is not affected, now the Vlasov surface kernels are structured similarly which should help with maintaining the Maxima.