Modify signature of applyBc and applyBcIdx functions and implement sheath BC in GkBasic
GK sheath boundary conditions require knowledge of the field, and as it was previously implemented we must also know the RK index so we pass the correct potentials to the BCs. For that reason this commit changes the signature of applyBc and applyBcIdx functions throughout the code, so we can pass `field`, `externalField`, `inIdx` and `outIdx`. It seems logical that one must pass `field` and `externalField` to the BCs, since the latter may depend on potentials and external magnetic fields. It's less clear that passing `inIdx` and `outIdx` is the best design, but seemed necessary for now. In the future it may even be necessary to pass the `species` table to BCs, e.g. for logical sheath BCs.

Since sheath BCs are almost fully contained within GkBasic now, we move sheath BC-related things that were in Equation/Gyrokinetic into App/BCs/GkBasic.lua, like phiWall and calcSheathReflection. We'd like to not use `GyrokineticModDecl` in App/BCs/GkBasic.lua too but we'll stick with it for now.

MomentSpecies regression tests pass except for those that require additional files or MPI.