Proposed changes to gkyl to deprecate the max order basis. This eliminates about 400 kernels from being compiled and on my Macbook reduced the build time from 15 minutes to 9.5 minutes.
Perhaps this branch should have a more general name, but I propose this branch be used as part of the discussion for streamlining kernel builds and especially to eliminate kernels which have no practical use (i.e., the max order basis).

It is not my current intention to eliminate things like the tensor basis, as the Vlasov side of things is currently exploring this extensively, especially because the tensor basis in Vlasov seems to have additional robustness and more easily allows the Vlasov solver to use upwind fluxes like the GK solver does.

That being said, this branch also eliminates p=1 Tensor basis kernels from being built where they were being built previously. We should have an indirection that if someone requests the tensor basis and p=1, it uses the Serendipity kernels since Serendipity p=1 and Tensor p=1 are identical bases.

Currently, this is just a proposal, and I have only eliminated the kernels from being compiled. Cleaning up of the headers, the indirection for p=1 Tensor, and deleting the kernels from the repo would need to be done before merging this back in.

I welcome comments.