Add support for STS in VmLBO
Add infrastructure in KineticSpecies and VmLBO to allow user to request STS stepping for VmLBO. This included adding a method and kernels in the equation object (Eq/VmLBO) which loops through the grid to find the minimum CFL frequency (a misnomer, since it uses the maximum speed/diffusion coefficient). Initial testing with Regression/vm-lbo/rt-lboRelax-1x1v-p1.lua suggests that gains of several factors in runtime can be obtained, especially as collision frequency and/or number of cells increase. More testing needed.

Some logic/program flow issues that need to be addressed:
- rk3opStep without the treatment="sts" in the collisions table doesn't appear
  to work.
- when one species uses STS but the other one doesn't, the program doesn't
  work.
I believe these two occur due to the way we are fetching CFL frequencies and setting number of stages in :sts.