Fix option to read source from file
This capability seems to have been lost when sources where moved into an App. Now one can do it two ways, through `fromFile=<file_name_suffix.bp>` or through `profile=<file_name_suffix.bp>`.

Note that there is some inconsistency in the code though. `fromFile` is meant to be used with `density` and `temperature`, so it calls MaxwellianProjection and DOES NOT scale by the phase-space Jacobian (B*). However if the file contains the raw source, without the phase-space Jacobian one should use `profile`, which then uses FunctionProjection and scales by that Jacobian. This should be made more consistent/error-prone.

Also, one has to be careful when importing sources and using `power` and `scaleWithSourcePower`. Because if you are importing a source that has already been scaled in another simulation, then it may think that the scaling factor for your ICs should just be 1, and you may start with the wrong ICs.