A few fixes in GkLogicalSheath and GkZeroCurrentField.
- Increase the tolerance in the GkLogicalSheath root finding. Might not be needed, but this is what E. Shi used.
- Implement partial moments for GK in DistFuncMomentCalc. Use these for computing boundary fluxes; the code appears more robust this way and is more reasonable (I think).
- Fix dz/2 factor in GkZeroCurrentField divP kernel.
- Remove an erroneous minus sign in GkZeroCurrentField.
- Some other fixes.

The ion acoustic Landau damping simulation still looks reasonable. However the mirror simualation is unstable. I have two hypotheses: a) just like with the E. Shi et al ELM pulse test, if the ICs are not sufficiently close to some steady state then the electrons react rapidly, the ions don't, quasineutrality is broken, and energy conservation is no longer preserved (and equation are inconsistent), or b) in the derivation of the field equation an IBP in vpar is used, does that suggest a surface term is needed in the discrete equations?