Fixed bug in the creation of vectors or MPI objects. Unfortunately, MPI is very horrible as there is a fundamental inconsistency in the way MPI manages objects. Sometimes it expects users to manage them, sometimes it manages them itself. This leads to a lot of confusion.
Also, the other problem is vector of FFI wrapper objects have to be
indexed in a very careful way. The semantics of vec[0] and vec+0 are
different. The former returns a value and the latter a pointer to the
value. Maybe there is a clean way to do this, but now the unit test
passes.