Partial progress from talking to Max and Mana (1,2) works but (2,1) doesnt
We believe it is an issue with contiguous indexing.
In the formal case the copy target is 6,7. In the broken case 5,8.
Our copy function, which uses localRange(), appears to mapping to 6,7 in
BOTH cases, which could explain why one works and the other doesn't.

We also learned that this bug is independent of decomposition cuts.
It persists for -n 1.